week5/week7/day1/exercise/students_root/requirements.txt
aiohttp==3.8.4
aiohttp-retry==2.8.3
aiosignal==1.3.1
asgiref==3.7.2
async-generator==1.10
async-timeout==4.0.2
attrs==23.1.0
backoff==2.2.1
beautifulsoup4==4.12.2
blinker==1.6.2
bs4==0.0.1
certifi==2023.5.7
chardet==3.0.4
charset-normalizer==3.1.0
click==8.1.3
defusedxml==0.7.1
distlib==0.3.6
Django==4.2.1
djangorestframework==3.14.0
exceptiongroup==1.1.1
Faker==18.7.0
filelock==3.12.0
Flask==2.3.2
frozenlist==1.3.3
googletrans==4.0.0rc1
h11==0.9.0
h2==3.2.0
hpack==3.0.0
hstspreload==2023.1.1
httpcore==0.9.1
httpie==3.2.2
httpx==0.13.3
hyperframe==5.2.0
idna==2.10
itsdangerous==2.1.2
Jinja2==3.1.2
joblib==1.2.0
libretranslatepy==2.1.1
lxml==4.9.2
markdown-it-py==2.2.0
MarkupSafe==2.1.2
mdurl==0.1.2
multidict==6.0.4
nltk==3.8.1
opencage==2.2.0
outcome==1.2.0
Pillow==9.5.0
platformdirs==3.5.1
psycopg2-binary==2.9.6
Pygments==2.15.1
PyJWT==2.7.0
PyPDF2==3.0.1
PySocks==1.7.1
python-dateutil==2.8.2
python-docx==0.8.11
python-dotenv==1.0.0
pytz==2023.3
regex==2023.5.5
requests==2.30.0
requests-toolbelt==1.0.0
rfc3986==1.5.0
rich==13.4.1
selenium==4.9.1
six==1.16.0
sniffio==1.3.0
sortedcontainers==2.4.0
soupsieve==2.4.1
sqlparse==0.4.4
tqdm==4.65.0
translate==3.6.1
trio==0.22.0
trio-websocket==0.10.2
twilio==8.2.1
urllib3==2.0.2
virtualenv==20.23.0
Werkzeug==2.3.4
wsproto==1.2.0
yarl==1.9.2
